backend/app/requirements.txt
flask==3.0.0
gunicorn==21.2.0
cryptography==41.0.7
flask-cors==4.0.0
psycopg2-binary==2.9.6
openai==1.3.0
requests==2.31.0